address PR review feedback from cgabriel
- Fix inconsistent iterable types in autoRestore (use [] instead of {})
- Move MockCache and MockApplication to common test module
- Replace print with logger.debug in multidb utility
