+++ gnrpy/tests/sql/test_gnrsql.py
@@ -5,33 +5,9 @@ import locale
 import pytest
 
 from gnr.sql import gnrsql as gs
-from .common import BaseGnrSqlTest
+from .common import BaseGnrSqlTest, MockApplication, MockCache
 
-class MockCache(object):
-    def __init__(self):
-        self._cache = {}
 
-    def getItem(self, key, defaultFactory=None):
-        if key not in self._cache and defaultFactory:
-            self._cache[key] = defaultFactory()
-        return self._cache.get(key)
-
-    def updatedItem(self, key):
-        if key in self._cache:
-            del self._cache[key]
-
-class MockApplication(object):
-    debug = True
-    config = { "db?reuse_relation_tree": False,
-               "db?auto_static_enabled": False,
-               'dbstores': None}
-    localizer = False
-    instanceFolder = None
-    cache = MockCache()
-
-    def checkResourcePermission(self, deletable, tags, test=True):
-        return test
-    
 class TestGnrSql(BaseGnrSqlTest):
 
     def test_generic_attrs(self):
